+++ src/backend/graphics/SDLRenderer2D.h
@@ -37,7 +37,7 @@ namespace Diamond {
 		/**
 		 Removes the render object at the given index from the rendering vector.
 		*/
-		void destroy_render_obj(int index);
+		void destroy_render_obj(unsigned long index);
 
 		/**
 		 Destroys window and renderer and shuts down SDL and SDL extensions.
