+++ src/modules/graphics/Graphics2D.cpp
@@ -22,6 +22,6 @@ void Diamond::Graphics2D::gen_render_obj(GameObject2D *parent, Texture *texture)
 	renderer->gen_render_obj(parent, texture);
 }
 
-void Diamond::Graphics2D::destroy_render_obj(int index) {
+void Diamond::Graphics2D::destroy_render_obj(unsigned long index) {
 	renderer->destroy_render_obj(index);
 }
